requirements.txt
django>=1.6.0
django-configurations>=0.7

# Optional dependencies
django-secure>=1.0
django_compressor>=1.3
django-crispy-forms>=1.4.0
django-log-request-id>=0.0.3
django-storages>=1.1.8
django-pylibmc-sasl>=0.2.4
django-heroku-memcacheify>=0.4
django-heroku-postgresify>=0.3
boto>=2.23.0
pusher>=0.8
raven>=4.0.3
structlog>=0.4.1
twilio
